Fixed for clarity
Added some user-friendly output such as specifying what keys need to be
added next. Also shuffled around the if/else statements so that all
error states are at the bottom. That's just for neatness sake.

Began work on the order_id portion. Currently decided NOT to include an
in-built QR scanner. Instead opted to have the user scan it themselves
and specify the order_id manually. Still a hell of a lot faster than
doing the whole process manually.